extract scriptDirectoryService out of executionZoneService
this breaks the circular dependency between executionZoneService and
scriptletBatchService.
